# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "XML Repair REST API",
    "version": "v1"
  },
  "paths": {
...
    "/api/repair": {
      "post": {
        "operationId": "repair",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairParams"
              }
            }
          },
          "description": "Repair parameters",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Processing completed successfully"
          }
        },
        "summary": "Repair XML issues in the specified entity"
      }
    },
    "/api/repairers": {
      "get": {
        "operationId": "listRepairers",
        "parameters": [
          {
            "description": "Entity type",
            "in": "query",
            "name": "entityType",
            "required": true,
            "schema": {
              "enum": [
                "COLLECTION",
                "DOCUMENT",
                "WORKITEM"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairerMeta"
                }
              }
            },
            "description": "Successfully retrieved the list of available repairers"
          }
        },
        "summary": "Get list of available repairers for the specified entity type"
      }
    },
    "/api/scan": {
      "post": {
        "operationId": "scan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanParams"
              }
            }
          },
          "description": "Scan parameters",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Scan completed successfully"
          }
        },
        "summary": "Scan a list of entities for XML issues based on the provided parameters"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RepairParams": {
        "properties": {
          "configs": {
            "additionalProperties": {
              "default": {},
              "description": "User custom repair configurations",
              "example": {
                "ModuleContentLinksRepairer": {
                  "convertToPlainText": true
                }
              },
              "type": "object"
            },
            "default": {},
            "description": "User custom repair configurations",
            "example": {
              "ModuleContentLinksRepairer": {
                "convertToPlainText": true
              }
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "issueMetaInfos": {
            "description": "List of issues meta infos to fix",
            "items": {
              "description": "List of issues meta infos to fix",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "issueMetaInfos"
        ],
        "type": "object"
      },
      "RepairerConfigMeta": {
        "properties": {
          "defaultValue": {
            "type": "object"
          },
          "description": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "type": {
            "enum": [
              "BOOLEAN"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
      "RepairerMeta": {
        "properties": {
          "configs": {
            "items": {
              "$ref": "#/components/schemas/RepairerConfigMeta"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ScanParams": {
        "properties": {
          "configs": {
            "additionalProperties": {
              "default": {},
              "description": "User custom repair configurations",
              "example": {
                "ModuleContentLinksRepairer": {
                  "convertToPlainText": true
                }
              },
              "type": "object"
            },
            "default": {},
            "description": "User custom repair configurations",
            "example": {
              "ModuleContentLinksRepairer": {
                "convertToPlainText": true
              }
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "entitySubtype": {
            "description": "Entity subtype to scan, e.g. 'requirement' for work items or 'specification' for documents (if not provided, all subtypes will be scanned)",
            "nullable": true,
            "type": "string"
          },
          "entityType": {
            "description": "Entity prototype to scan. Currently supported values are: WORKITEM, DOCUMENT and COLLECTION",
            "enum": [
              "COLLECTION",
              "DOCUMENT",
              "WORKITEM"
            ],
            "example": "WORKITEM",
            "type": "string"
          },
          "hideValid": {
            "default": false,
            "description": "Hide entities without issues from the resulting list",
            "nullable": true,
            "type": "boolean"
          },
          "limit": {
            "default": 100,
            "description": "Maximum number of entities to scan",
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "projectId": {
            "description": "The unique identifier for the project",
            "example": "elibrary",
            "type": "string"
          },
          "repairers": {
            "description": "List of repairers to use (if no data provided the default repairers will be applied)",
            "items": {
              "description": "List of repairers to use (if no data provided the default repairers will be applied)",
              "nullable": true,
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "sort": {
            "default": "created",
            "description": "Sorting criteria in the format 'created' or '~id'",
            "nullable": true,
            "type": "string"
          },
          "timeout": {
            "default": 60000,
            "description": "Operation timeout, milliseconds",
            "format": "int64",
            "type": "integer"
          },
          "userQuery": {
            "description": "Additional query to select entities for scanning, e.g. 'id:TEST-12345'",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  }
}